Create README.md
This was just a file in which I finally found out how to create working code and I just did a bunch of stuff in it messing around because I am self Taught and I don't know what to and what not to do so. This was just me messing around. Because I finally figured out how to write working code and get it to do stuff that I wanted to do.
So all of the code in it. I am not. Being held responsible for because I just made stuff up as I went. But it is working code and I have learned and it has taught me a lot. If any problems arise from me doing this, then this whole read me an any other read means. I Have are considered from that point on null and void for everything I did and all My rights to or over anything will terminate, and so will the copyrights They also will terminate and be considered Null and void but if anyone wants to use any part of it free of charge and you are granted everything you need to change my code in any way I cannot and will not agree to be held responsible for any liability legal responsibilities and I do not Guarantee this code in any way if you use it. It is on an as is the basis and you're taking the risk yourself. As I did to make it.
And this applies to every. Enter the or being or person. Use at your own risk. What is what I'm saying? And if you try to come after me for anything it all will be considered terminated. An all Terms & Conditions will be terminated. All copyright will be terminated. All permissions will be terminated And anything I said about you using This software is terminated and considered off the table. But if you want to mess around with it, go for it. But don't hold me responsible for any reason. A rising from any of the software.